fix: close nine independent-inspection findings, and verify against real Postgres
claudex-loop ran over the whole build for the first time (codex/gpt-6-astra,
xhigh): REVISE, 38 findings. Nine are fixed here, each with a test; all 38 are
recorded with a plan in IR-FINDINGS.md.

IR-002 (high) — one unauthenticated request killed the process. A malformed
escape threw URIError out of the HTTP listener. Reproduced against a running
server, then fixed: the router reports a malformed path and the listener body
is guarded.

IR-003 (high) — the embedded driver is one connection shared by every request,
so BEGIN/COMMIT as separate statements let two requests interleave and one
rollback discard another's committed rows. Now delegates to PGlite's exclusive
transaction. Tested with 20 concurrent transactions, half rolling back.

IR-006/IR-007 (high) — the SMTP timeout was never stored, so STARTTLS disabled
it; and the AUTH commands were not redacted, so a server echoing the credential
leaked it into an error that is persisted and logged.

IR-011 (high) — the README documented TRANSLATE_API_URL while the loader read
TRANSLATE_BASE_URL, so following the docs sent bug text and the API key to
api.openai.com. Both names now work.

IR-012 (high) — production defaulted to the public dev HMAC secret and
non-Secure cookies. Production now refuses to start without STORAGE_SECRET.

IR-014/IR-024/IR-033 — migrations no longer race across replicas
(MIGRATE_ON_START); readiness emails carry the application URL instead of a
placeholder; the container binds 0.0.0.0 so the published port works.

Also: the SQL had only ever run on PGlite. Stood up a real PostgreSQL 17.11 and
ran the whole suite against it — which immediately caught a count(*) that
node-postgres returns as a string, so the bug list sent "3" instead of 3.

Suite: 251 passing on the embedded driver, through the pg driver, and against
real PostgreSQL.

+++ src/api.js
@@ -475,7 +475,10 @@ export function buildRoutes() {
     const rows = await ctx.db.query(
       `SELECT b.id, b.bug_number, b.severity, b.status, b.title_vi, b.milestone_id,
               b.updated_at, m.code AS milestone_code,
-              (SELECT count(*) FROM bug_attachments a WHERE a.bug_id = b.id) AS attachments
+              -- ::int matters: count() is int8, which node-postgres returns as a
+              -- *string* (a JS number cannot hold every int8). Without the cast the
+              -- row would carry "3" on PostgreSQL and 3 on PGlite.
+              (SELECT count(*) FROM bug_attachments a WHERE a.bug_id = b.id)::int AS attachments
          FROM bugs b JOIN milestones m ON m.id = b.milestone_id
         WHERE b.project_id = $1 AND b.deleted_at IS NULL
         ORDER BY b.bug_number DESC`, [ctx.params.id]);
